+++ internal/config/config.go
@@ -1,0 +1,9 @@
+package config
+
+const (
+	Host = "dev"
+	Port = 2223
+
+	ChallengeDir = "./chals"
+	DownloadRoot = "./downloads"
+)
