--- util.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"log"
-	"math/rand"
-	"runtime/debug"
-)
-
-func must(err error) {
-	if err != nil {
-		debug.PrintStack()
-		log.Fatal(err)
-	}
-}
-
-func randHex(n int) string {
-	const letters = "0123456789abcdef"
-	result := make([]byte, n)
-	for i := range result {
-		result[i] = letters[rand.Intn(len(letters))]
-	}
-	return string(result)
-}
